Added to_dict() for exporting metadata
Currently there is no way of exporting only the metadata as a dictionary. There are ways of obtaining the values alone or the keys alone, or items as a list of tuples. But if I want a dict to use it externally as a dict I would have to use metadata.keys() and iterate over the keys in order to get the values for each key and add each key-value to a new dict, which seems wasteful, since there is already an internal field of type dict holding the info I am interested in.